# Improve performance of example-nextjs by utilizing babel-loader