--- a/examples/example-nextjs/next.config.js
+++ b/examples/example-nextjs/next.config.js
@@ -11,4 +11,5 @@
 
 module.exports = {
   transpilePackages: ['@stylexjs/open-props'],
+  eslint: { ignoreDuringBuilds: true },
 };