--- a/examples/example-nextjs/next.config.js
+++ b/examples/example-nextjs/next.config.js
@@ -9,7 +9,45 @@
 
 /** @type {import('next').NextConfig} */
 
+const path = require('path');
+
 module.exports = {
   transpilePackages: ['@stylexjs/open-props'],
   eslint: { ignoreDuringBuilds: true },
+  webpack: (config, { dev, isServer }) => {
+    // Process only files that require StyleX compilation using babel-loader
+    config.module.rules.push({
+      test: /\.(tsx|stylex\.ts|jsx|stylex\.js)$/,
+      exclude: /node_modules(?!\/@stylexjs\/open-props)/,
+      use: [
+        {
+          loader: 'babel-loader',
+          options: {
+            parserOpts: {
+              plugins: ['typescript', 'jsx'],
+            },
+            plugins: [
+              [
+                '@stylexjs/babel-plugin',
+                {
+                  dev: dev,
+                  runtimeInjection: false,
+                  genConditionalClasses: true,
+                  treeshakeCompensation: true,
+                  aliases: {
+                    '@/*': [path.join(__dirname, '*')],
+                  },
+                  unstable_moduleResolution: {
+                    type: 'commonJS',
+                  },
+                },
+              ],
+            ],
+          },
+        },
+      ],
+    });
+
+    return config;
+  },
 };