--- a/examples/example-nextjs/package.json
+++ b/examples/example-nextjs/package.json
@@ -10,19 +10,20 @@
     "example:start": "next start"
   },
   "dependencies": {
-    "@stylexjs/stylex": "0.11.1",
     "@stylexjs/open-props": "0.11.1",
+    "@stylexjs/stylex": "0.11.1",
     "next": "14.2.21",
     "react": "^18.3.0",
-    "react-dom": "^18.3.0"
+    "react-dom": "^18.3.0",
+    "styled-jsx": "^5.1.1"
   },
   "devDependencies": {
     "@stylexjs/eslint-plugin": "0.11.1",
     "@stylexjs/postcss-plugin": "0.11.1",
+    "@types/json-schema": "^7.0.15",
     "@types/node": "^22.7.6",
     "@types/react": "^18.3.0",
     "@types/react-dom": "^18.3.0",
-    "@types/json-schema": "^7.0.15",
     "@typescript-eslint/parser": "^6.21.0",
     "eslint": "^8.57.1",
     "eslint-config-next": "14.0.1",