… examples/example-nextjs/babel.config.js