## What changed / motivation ?

This PR improves the performance of the `example-nextjs` project by implementing
`babel-loader`. Utilizing `babel-loader` for JavaScript/TypeScript transpilation
leads to faster build times and an improved development experience compared to
the previous setup.

## Additional Context

**Technical Details:**

- Using `babel.config.js` previously forced all compilation through Babel,
  slowing down the build process.
- This change configures `babel-loader` within the webpack configuration
  (`next.config.js`). This ensures only files requiring StyleX compilation are
  processed by Babel, while the rest leverage Next.js's faster SWC compiler.
- Additionally, the PostCSS compilation target has been narrowed down to only
  `.tsx` and `.stylex.ts` files, further optimizing the build pipeline.

**Benchmark Results:**

_Command: `npm run -w ./examples/example-nextjs example:build`_

**Before:**

```
Time (mean ± σ):     14.754 s ±  0.343 s    [User: 24.845 s, System: 2.581 s]
Range (min … max):   14.385 s … 15.173 s    5 runs
```

**After:**

```
Time (mean ± σ):     11.423 s ±  0.395 s    [User: 19.161 s, System: 2.293 s]
Range (min … max):   10.918 s … 11.841 s    5 runs
```

**Internal Codebase Improvement:** Furthermore, we observed a significant build
time improvement in our internal company codebase, reducing from 397.48s to
271.48s after applying these changes.

**Other Changes:**

- The `example:build` script in `example-nextjs` was failing. Added `styled-jsx`
  to `examples/example-nextjs/package.json` as a dependency to resolve this
  issue.
- Due to a large number of pre-existing lint errors, linting during the Next.js
  build process has been temporarily disabled
  (`eslint: { ignoreDuringBuilds: true }` in `next.config.js`) to allow the
  build to complete successfully. Addressing these lint errors is outside the
  scope of this performance-focused PR.

**Acknowledgements:** Special thanks to my mentor @javascripter for their
valuable advice and guidance on implementing these performance improvements.

## Pre-flight checklist

- [x] I have read the contributing guidelines
    [Contribution Guidelines](https://github.com/facebook/stylex/blob/main/.github/CONTRIBUTING.md)
- [x] Performed a self-review of my code