@javascripter 2025-03-30
For context:

Using`babel.config.js` forces Next.js to compile a lot of files that don't need to be transformed with StyleX. Using `babel-loader` allows you to selectively compile files with StyleX Babel Plugin while delegating most files to Next.js default config.

For example, if we print the path of processed files in StyleX Babel Plugin, we see lots of internal files that don't need to be processed by StyleX Babel Plugin in the current main branch.
```
   Disabled SWC as replacement for Babel because of custom Babel configuration "babel.config.js" https://nextjs.org/docs/messages/swc-disabled
 ✓ Ready in 1378ms
   Using external babel configuration from /Users/tsubasa/Projects/github.com/facebook/stylex/examples/example-nextjs/babel.config.js
StyleX: path /Users/tsubasa/Projects/github.com/facebook/stylex/examples/example-nextjs/app/layout.tsx
StyleX: path /Users/tsubasa/Projects/github.com/facebook/stylex/examples/example-nextjs/app/page.tsx
StyleX: path /Users/tsubasa/Projects/github.com/facebook/stylex/examples/example-nextjs/node_modules/next/dist/server/future/route-kind.js
StyleX: path /Users/tsubasa/Projects/github.com/facebook/stylex/examples/example-nextjs/node_modules/next/dist/server/app-render/entry-base.js
StyleX: path /Users/tsubasa/Projects/github.com/facebook/stylex/examples/example-nextjs/node_modules/next/dist/client/components/error-boundary.js
StyleX: path /Users/tsubasa/Projects/github.com/facebook/stylex/examples/example-nextjs/node_modules/next/dist/client/components/not-found-error.js
StyleX: path /Users/tsubasa/Projects/github.com/facebook/stylex/examples/example-nextjs/node_modules/next/dist/server/future/route-modules/app-page/module.compiled.js
 ○ Compiling / ...
```

Next.js itself [uses](https://github.com/vercel/next.js/blob/3261044fcb3b0fb96c021f4e0ba0b7a7c25b5b77/packages/next/src/build/get-babel-loader-config.ts#L40) babel-loader to run React Compiler with babel as well and this PR is also is compatible with their [`experimental.reactCompiler` option](https://nextjs.org/docs/app/api-reference/config/next-config-js/reactCompiler).