@BMCwebdev 2025-06-06
I know this PR is closed, but hoping I could still ask @javascripter and @km-tr a question in here. I love the work done here, and I have implemented it, but I run into an issue. Trying to get Jest to run without a `babel.config.js` I can't get Jest to compile the StyleX code. I also have a component library that is built with uncompiled StyleX code. So in consuming apps, this library is a packages in node_modules. This setup you have above works great for building and dev, but again, I cannot get Jest to work this way. I've tried creating a jest-transformer and jest.config to try and use the same setup, but still doesn't work

```
  The Jest Problem:
  The exact same StyleX configuration works in Next.js webpack but fails in
  Jest. Since the code is identical, this must be a Jest-specific module 
  resolution or processing issue.

  Key Clue:
  The error shows Jest referencing:
  /src/components/Theme/dynamicThemeUtils.jsx:14:42

  But we confirmed this path doesn't exist in the published package. This
  suggests Jest is either:

  1. Following source maps to non-existent source files
  2. Processing internal imports within the bundled dist files differently than
   webpack
  3. Using different module resolution with pnpm's symlink structure
  ```