@javascripter 2025-06-06
@BMCwebdev 
Have you configured jest with babel-jest to transform files with babel? I haven't verified the below but in my experience you need to follow these steps:

1. Update [transformIgnorePatterns](https://jestjs.io/docs/configuration#transformignorepatterns-arraystring) to make sure your component library gets transformed via babel-jest.
2. Create babel.config.js and verify your tests pass. This confirms your babel config and `transformIgnorePatterns` are configured properly and is working as expected.
3. Either rename babel.config.js to custom-babel.config.js or inline your babel config there:

https://www.npmjs.com/package/babel-jest
```
// jest.config.js
"transform": {
  "\\.[jt]sx?$": [
    "babel-jest",
    {
      "extends": "./custom-babel.config.js" // set your custom babel config path
       // or inline your babel config here
    }
  ]
},
```